Former-commit-id: aecc740a518bca45da58c550b1b829e80e50746e

+++ framework/sofa/core/behavior/MechanicalState.h
@@ -84,8 +84,8 @@ public:
 protected:
     virtual ~MechanicalState() {}
 public:
-    virtual unsigned int getCoordDimension() const { return (unsigned int)defaulttype::DataTypeInfo<Coord>::size(); }
-    virtual unsigned int getDerivDimension() const { return (unsigned int)defaulttype::DataTypeInfo<Deriv>::size(); }
+    virtual size_t getCoordDimension() const { return defaulttype::DataTypeInfo<Coord>::size(); }
+    virtual size_t getDerivDimension() const { return defaulttype::DataTypeInfo<Deriv>::size(); }
 
     /// Get the indices of the particles located in the given bounding box
     virtual void getIndicesInSpace(sofa::helper::vector<unsigned>& /*indices*/, Real /*xmin*/, Real /*xmax*/,Real /*ymin*/, Real /*ymax*/, Real /*zmin*/, Real /*zmax*/) const=0;
