+++ framework/sofa/core/objectmodel/BaseLink.h
@@ -127,7 +127,7 @@ public:
     /// This can be used to efficiently detect changes
     int getCounter(const core::ExecParams* params) const { return m_counters[core::ExecParams::currentAspect(params)]; }
 
-    virtual unsigned int getSize() const = 0;
+    virtual size_t getSize() const = 0;
     virtual Base* getLinkedBase(unsigned int index=0) const = 0;
     virtual BaseData* getLinkedData(unsigned int index=0) const = 0;
     virtual std::string getLinkedPath(unsigned int index=0) const = 0;
